1401.04.09

+++ app/Http/Controllers/StoriesController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\CreateStoryReq;
+use App\Http\Requests\DeleteStoryReq;
 use Illuminate\Http\Request;
 use App\Models\Stories;
 use Illuminate\Support\Str;
@@ -14,25 +16,21 @@ use Hekmatinasser\Verta\Verta;
 
 class StoriesController extends Controller
 {
-    public function CreateStory(request $req)
+    public function CreateStory(request $req, CreateStoryReq $valid)
     {
-        if (!empty($req->file('input'))) {
+        if (!empty($valid->file('input'))) {
 
-            $fileName = time() . '_' . $req->file('input')->getClientOriginalName();
-            $checkupload = Storage::disk('sv')->put('stories', $req->file('input'));
+            $fileName = time() . '_' . $valid->file('input')->getClientOriginalName();
+            $checkupload = Storage::disk('sv')->put('stories', $valid->file('input'));
 
             $db = Stories::create([
                 'user_id' => $req->user()->id,
                 'input' => $checkupload,
             ]);
-            $isdone = Controller::isDone($db);
-            return response()->json([
-                'isDone' => $isdone,
-            ]);
+
+            return Controller::Response('', true, 'created');
         } else {
-            return response()->json([
-                'isDone' => false,
-            ]);
+            return Controller::Response('', false, 'some error');
         }
     }
 
@@ -76,7 +74,7 @@ class StoriesController extends Controller
                 }
             }
         }
-        $newmystory[] = ['name' => $req->user()->name, 'img' => env('DEFAULT_URL') . $req->user()->img, 'stories' =>  $sd];
+        $newmystory[] = ['user_id' => $req->user()->id, 'name' => $req->user()->name, 'img' => env('DEFAULT_URL') . $req->user()->img, 'stories' =>  $sd];
 
         $stories = FollowingPosts::where('following_id', $req->user()->id)->with('stories')->get()->makehidden(['id', 'following_id', 'status', 'date'])->toArray();
 
@@ -117,6 +115,7 @@ class StoriesController extends Controller
                 $last_story = end($sm);
                 $last_story_date = strtotime($last_story['created_at']);
                 $str[] = [
+                    'user_id' => $usr->id,
                     'name' => $usr->name,
                     'img' => env('DEFAULT_URL') . $usr->img,
                     'stories' =>  $sm,
@@ -151,13 +150,14 @@ class StoriesController extends Controller
             'data' => $db
         ]);
     }
-    public function DeleteStory(request $req)
+    public function DeleteStory(request $req, DeleteStoryReq $valid)
     {
-        $story = Stories::where('id', $req->id)->delete();
-        $story_views = StoryViews::where('story_id', $req->id)->delete();
-        return response()->json([
-            'isDone' => true,
-            'data' => 'done'
-        ]);
+        $story = Stories::where('id', $valid->id)->where('user_id', $req->user()->id)->delete();
+        if ($story) {
+            $story_views = StoryViews::where('story_id', $valid->id)->delete();
+            return Controller::Response('', true, 'deleted');
+        } else {
+            return Controller::Response('', false, 'some error');
+        }
     }
 }
